requirements.txt
apiclient==1.0.3
appnope==0.1.0
asn1crypto==0.24.0
azure-common==1.1.11
azure-nspkg==2.0.0
azure-storage==0.36.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.4
boto==2.48.0
boto3==1.5.0
botocore==1.8.14
branca==0.3.0
bs4==0.0.1
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
coverage==4.5.1
cryptography==2.2.2
cycler==0.10.0
datadog==0.21.0
DateTime==4.2
decorator==4.3.0
defusedxml==0.5.0
descartes==1.1.0
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
Fiona==1.7.11.post2
fire==0.1.3
folium==0.5.0
future==0.16.0
geojson==2.4.0
geopandas==0.3.0
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
gspread==0.6.2
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
ijson==2.3
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
kml2geojson==4.0.2
Lifetimes==0.9.0.0
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.14.3
oauth2client==4.1.2
opencage==1.1.4
pandas==0.23.0
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
plotly==2.7.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodome==3.6.1
PyDrive==1.3.1
pygeohash==1.2.0
Pygments==2.2.0
PyJWT==1.6.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyproj==1.9.5.1
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.18.4
rsa==3.4.2
Rtree==0.8.3
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
sklearn==0.0
slack-cleaner==0.3.0
slackclient==1.2.1
slacker==0.9.65
snowflake-connector-python==1.6.1
snowflake-sqlalchemy==1.1.1
SQLAlchemy==1.2.8
statsmodels==0.8.0
tabulate==0.8.2
terminado==0.8.1
testpath==0.3.1
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
widgetsnbextension==3.4.2
zope.interface==4.5.0
